+++ frontend/public/components/app.jsx
@@ -19,6 +19,7 @@ import { Masthead } from './masthead';
 import { getBrandingDetails } from './utils/branding';
 import { ConsoleNotifier } from './console-notifier';
 import { NotificationDrawer } from './notification-drawer';
+import { NotificationDrawer } from './notification-drawer';
 import { Navigation } from '@console/app/src/components/nav';
 import { history, AsyncComponent, LoadingBox, useSafeFetch, usePoll } from './utils';
 import * as UIActions from '../actions/ui';
@@ -74,6 +75,7 @@ import { graphQLReady } from '../graphql/client';
 import { AdmissionWebhookWarningNotifications } from '@console/app/src/components/admission-webhook-warnings/AdmissionWebhookWarningNotifications';
 import { usePackageManifestCheck } from '@console/shared/src/hooks/usePackageManifestCheck';
 import { useCSPViolationDetector } from '@console/app/src/hooks/useCSPVioliationDetector';
+import { useNotificationPoller } from '@console/app/src/hooks/useNotificationPoller';
 
 initI18n();
 
@@ -110,16 +112,13 @@ const App = (props) => {
 
   const [isMastheadStacked, setIsMastheadStacked] = React.useState(isMobile());
   const [isNavOpen, setIsNavOpen] = React.useState(isDesktop());
-  const [isDrawerInline, setIsDrawerInline] = React.useState(isLargeLayout());
 
   const previousDesktopState = React.useRef(isDesktop());
   const previousMobileState = React.useRef(isMobile());
-  const previousDrawerInlineState = React.useRef(isLargeLayout());
 
   const onResize = React.useCallback(() => {
     const desktop = isDesktop();
     const mobile = isMobile();
-    const drawerInline = isLargeLayout();
     if (previousDesktopState.current !== desktop) {
       setIsNavOpen(desktop);
       previousDesktopState.current = desktop;
@@ -128,13 +127,10 @@ const App = (props) => {
       setIsMastheadStacked(mobile);
       previousMobileState.current = mobile;
     }
-    if (previousDrawerInlineState.current !== drawerInline) {
-      setIsDrawerInline(drawerInline);
-      previousDrawerInlineState.current = drawerInline;
-    }
   }, []);
 
   useCSPViolationDetector();
+  useNotificationPoller();
 
   React.useEffect(() => {
     window.addEventListener('resize', onResize);
@@ -217,6 +213,10 @@ const App = (props) => {
 
   const { productName } = getBrandingDetails();
 
+  const isNotificationDrawerExpanded = useSelector(
+    (state) => !!state.UI.getIn(['notifications', 'isExpanded']),
+  );
+
   const content = (
     <>
       <Helmet titleTemplate={`%s · ${productName}`} defaultTitle={productName} />
@@ -247,13 +247,15 @@ const App = (props) => {
                 Skip to Content
               </SkipToContent>
             }
+            notificationDrawer={
+              <NotificationDrawer
+                onDrawerChange={onNotificationDrawerToggle}
+                isDrawerExpanded={isNotificationDrawerExpanded}
+              />
+            }
+            isNotificationDrawerExpanded={isNotificationDrawerExpanded}
           >
-            <NotificationDrawer
-              isDesktop={isDrawerInline}
-              onDrawerChange={onNotificationDrawerToggle}
-            >
-              <AppContents />
-            </NotificationDrawer>
+            <AppContents />
           </Page>
           {consoleCapabilityLightspeedButtonIsEnabled && lightspeedIsAvailableToInstall && (
             <Lightspeed />
